Terminando caso de uso Registro, enlazado al login
