+++ src/main/resources/templates/listar.html
@@ -28,19 +28,21 @@
             ><img src="https://raw.githubusercontent.com/MartinHeinz/MartinHeinz/master/wave.gif" width="30px" />
           </h1>
           <hr class="my-4" />
-          <h6 sec:authorize="hasRole('ADMIN')">
+          <h6 sec:authorize="hasAuthority('ADMIN')">
             En este apartado podrá gestionar todo lo referente a los jueces, podrá registrar nuevos jueces, editar su información,
             eliminarlos o ver mas detalles.
           </h6>
-          <h6 sec:authorize="hasRole('USER')">En este apartado podrá consultar información detallada de los jueces registrados</h6>
-          <h6 sec:authorize="hasRole('ENTRENADOR')">En este apartado podrá consultar información detallada de los jueces registrados</h6>
+          <h6 sec:authorize="hasAuthority('JUEZ')">En este apartado podrá consultar información detallada de los jueces registrados</h6>
+          <h6 sec:authorize="hasAuthority('ENTRENADOR')">
+            En este apartado podrá consultar información detallada de los jueces registrados
+          </h6>
         </div>
         <div class="card text-white" style="background-color: #ffffff2c">
           <h4 class="card-header text-white" style="background-color: #001f61" th:text="${titulo}"></h4>
 
           <div class="card-body table-responsive">
             <h4 class="card-title">
-              <a th:href="@{/form}" class="btn btn-primary" sec:authorize="hasRole('ADMIN')"
+              <a th:href="@{/form}" class="btn btn-primary" sec:authorize="hasAuthority('ADMIN')"
                 ><i class="fa-solid fa-circle-plus"></i> Agregar juez</a
               >
             </h4>
@@ -57,8 +59,8 @@
                   <th>Edad</th>
                   <th>Foto</th>
                   <th>Detalles</th>
-                  <th sec:authorize="hasRole('ADMIN')">Editar</th>
-                  <th sec:authorize="hasRole('ADMIN')">Eliminar</th>
+                  <th sec:authorize="hasAuthority('ADMIN')">Editar</th>
+                  <th sec:authorize="hasAuthority('ADMIN')">Eliminar</th>
                 </tr>
               </thead>
               <tbody>
@@ -74,12 +76,12 @@
                   <td>
                     <a class="btn btn-info btn-sm" th:href="@{/ver/} + ${juez.id}"><i class="fa-solid fa-circle-info"></i> Detalles</a>
                   </td>
-                  <td sec:authorize="hasRole('ADMIN')">
+                  <td sec:authorize="hasAuthority('ADMIN')">
                     <a class="btn btn-warning btn-sm text-white" th:href="@{/form/} + ${juez.id}"
                       ><i class="fa-solid fa-pen-to-square"></i> Editar</a
                     >
                   </td>
-                  <td sec:authorize="hasRole('ADMIN')">
+                  <td sec:authorize="hasAuthority('ADMIN')">
                     <a class="btn btn-danger btn-sm text-white" th:onclick="eliminar([[${juez.id}]])"
                       ><i class="fa-solid fa-trash-can"></i> Eliminar</a
                     >
