+++ src/main/resources/templates/listarCompetidores.html
@@ -27,18 +27,22 @@
             ><img src="https://raw.githubusercontent.com/MartinHeinz/MartinHeinz/master/wave.gif" width="30px" />
           </h1>
           <hr class="my-4" />
-          <h6 sec:authorize="hasRole('ENTRENADOR')">
+          <h6 sec:authorize="hasAuthority('ENTRENADOR')">
             En este apartado podrá gestionar todo lo referente a los competidores, podrá registrar nuevos competidores, editar su
             información, eliminarlos o ver mas detalles.
           </h6>
-          <h6 sec:authorize="hasRole('USER')">En este apartado podrá consultar información detallada de los competidores registrados</h6>
-          <h6 sec:authorize="hasRole('ADMIN')">En este apartado podrá consultar información detallada de los competidores registrados</h6>
+          <h6 sec:authorize="hasAuthority('JUEZ')">
+            En este apartado podrá consultar información detallada de los competidores registrados
+          </h6>
+          <h6 sec:authorize="hasAuthority('ADMIN')">
+            En este apartado podrá consultar información detallada de los competidores registrados
+          </h6>
         </div>
         <div class="card text-white" style="background-color: #ffffff2c">
           <h4 class="card-header text-white" style="background-color: #001f61" th:text="${titulo}"></h4>
           <div class="card-body table-responsive">
             <h4 class="card-title">
-              <a th:href="@{/crear_competidor}" class="btn btn-primary" sec:authorize="hasRole('ENTRENADOR')"
+              <a th:href="@{/crear_competidor}" class="btn btn-primary" sec:authorize="hasAuthority('ENTRENADOR')"
                 ><i class="fa-solid fa-circle-plus"></i> Agregar Competidor</a
               >
             </h4>
@@ -68,8 +72,8 @@
                   <th>Edad</th>
                   <th>Foto</th>
                   <th>Detalles</th>
-                  <th sec:authorize="hasRole('ENTRENADOR')">Editar</th>
-                  <th sec:authorize="hasRole('ENTRENADOR')">Eliminar</th>
+                  <th sec:authorize="hasAuthority('ENTRENADOR')">Editar</th>
+                  <th sec:authorize="hasAuthority('ENTRENADOR')">Eliminar</th>
                 </tr>
               </thead>
               <tbody>
@@ -88,12 +92,12 @@
                       ><i class="fa-solid fa-circle-info"></i> Detalles</a
                     >
                   </td>
-                  <td sec:authorize="hasRole('ENTRENADOR')">
+                  <td sec:authorize="hasAuthority('ENTRENADOR')">
                     <a class="btn btn-warning btn-sm text-white" th:href="@{/crear_competidor/} + ${competidorPorDisciplina.id}"
                       ><i class="fa-solid fa-pen-to-square"></i> Editar</a
                     >
                   </td>
-                  <td sec:authorize="hasRole('ENTRENADOR')">
+                  <td sec:authorize="hasAuthority('ENTRENADOR')">
                     <a class="btn btn-danger btn-sm text-white" th:onclick="eliminar([[${competidorPorDisciplina.id}]])"
                       ><i class="fa-solid fa-trash-can"></i> Eliminar</a
                     >
