+++ src/main/resources/templates/listarDisciplinas.html
@@ -26,12 +26,14 @@
             ><img src="https://raw.githubusercontent.com/MartinHeinz/MartinHeinz/master/wave.gif" width="30px" />
           </h1>
           <hr class="my-4" />
-          <h6 sec:authorize="hasRole('ADMIN')">
+          <h6 sec:authorize="hasAuthority('ADMIN')">
             En este apartado podrá gestionar todo lo referente a las disciplinas, podrá registrar nuevas disciplinas, editar su información,
             eliminarlas o ver mas detalles.
           </h6>
-          <h6 sec:authorize="hasRole('USER')">En este apartado podrá consultar información detallada de las disciplinas registradas</h6>
-          <h6 sec:authorize="hasRole('ENTRENADOR')">
+          <h6 sec:authorize="hasAuthority('JUEZ')">
+            En este apartado podrá consultar información detallada de las disciplinas registradas
+          </h6>
+          <h6 sec:authorize="hasAuthority('ENTRENADOR')">
             En este apartado podrá consultar información detallada de las disciplinas registradas
           </h6>
         </div>
@@ -39,7 +41,7 @@
           <h4 class="card-header text-white" style="background-color: #001f61" th:text="${titulo}"></h4>
           <div class="card-body table-responsive">
             <h4 class="card-title">
-              <a th:href="@{/crear_disciplina}" class="btn btn-primary" sec:authorize="hasRole('ADMIN')"
+              <a th:href="@{/crear_disciplina}" class="btn btn-primary" sec:authorize="hasAuthority('ADMIN')"
                 ><i class="fa-solid fa-circle-plus"></i> Agregar disciplina</a
               >
             </h4>
@@ -50,8 +52,8 @@
                   <th>#</th>
                   <th>Disciplina</th>
                   <th>Detalles</th>
-                  <th sec:authorize="hasRole('ADMIN')">Editar</th>
-                  <th sec:authorize="hasRole('ADMIN')">Eliminar</th>
+                  <th sec:authorize="hasAuthority('ADMIN')">Editar</th>
+                  <th sec:authorize="hasAuthority('ADMIN')">Eliminar</th>
                 </tr>
               </thead>
               <tbody>
@@ -63,12 +65,12 @@
                       ><i class="fa-solid fa-circle-info"></i> Detalles</a
                     >
                   </td>
-                  <td sec:authorize="hasRole('ADMIN')">
+                  <td sec:authorize="hasAuthority('ADMIN')">
                     <a class="btn btn-warning btn-sm text-white" th:href="@{/crear_disciplina/} + ${disciplina.id}"
                       ><i class="fa-solid fa-pen-to-square"></i> Editar</a
                     >
                   </td>
-                  <td sec:authorize="hasRole('ADMIN')">
+                  <td sec:authorize="hasAuthority('ADMIN')">
                     <a class="btn btn-danger btn-sm text-white" th:onclick="eliminar([[${disciplina.id}]])"
                       ><i class="fa-solid fa-trash-can"></i> Eliminar</a
                     >
