+++ src/main/resources/templates/registro.html
@@ -3,7 +3,7 @@
 <html xmlns:th="http://www.thymeleaf.org">
   <head th:fragment="head">
     <meta charset="UTF-8" />
-    <title>Login</title>
+    <title>Registro Entrenadores</title>
     <!-- CSS only -->
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
@@ -53,7 +53,7 @@
                   alt="Prometeo"
                 />
                 <span>| Prometeo <i class="fa-solid fa-fire" style="color: #ff7500"></i></span>
-                <h3 class="mb-5">Inicio de sesión</h3>
+                <h3 class="mb-5">Registro Entrenadores</h3>
                 <div th:if="${param.error}">
                   <div class="alert alert-danger alert-dismissible fade show" role="alert">
                     Username o password inválidos
@@ -64,23 +64,54 @@
                 </div>
 
                 <div th:if="${param.logout}">
-                  <div class="alert alert-primary alert-dismissible fade show" role="alert">
-                    Ha cerrado sesión exitosamente
-                    <button type="button" class="btn-close" data-dismiss="alert" aria-label="Close"></button>
+                  <div th:if="${param.exito}">
+                    <div class="alert alert-info">Se ha registrado exitosamente a la aplicación</div>
                   </div>
                 </div>
-                <form th:action="@{/login}" method="post">
-                  <div class="form-outline mb-4">
-                    <input type="text" id="username" name="username" class="form-control form-control-lg" />
-                    <label class="form-label" for="typeEmailX-2">Nombre de usuario</label>
+                <form th:action="@{/registro}" method="post" th:object="${usuario}">
+                  <div class="form-group mb-1">
+                    <label class="control-label" for="nombre">Nombre : </label>
+                    <input id="nombre" type="text" class="form-control" th:field="*{nombre}" required autofocus="autofocus" />
                   </div>
 
-                  <div class="form-outline mb-4">
-                    <input type="password" id="password" name="password" class="form-control form-control-lg" />
-                    <label class="form-label" for="typePasswordX-2">Password</label>
+                  <div class="form-group mb-1">
+                    <label class="control-label" for="apellidoPaterno">Apellido Paterno : </label>
+                    <input id="apellidoPaterno" class="form-control" th:field="*{apellidoPaterno}" required autofocus="autofocus" />
                   </div>
 
-                  <button class="btn btn-primary btn-lg btn-block" type="submit">Iniciar sesión</button>
+                  <div class="form-group mb-1">
+                    <label class="control-label" for="apellidoMaterno">Apellido Materno : </label>
+                    <input id="apellidoMaterno" class="form-control" th:field="*{apellidoMaterno}" required autofocus="autofocus" />
+                  </div>
+
+                  <div class="form-group mb-1">
+                    <label for="edad">Edad:</label>
+                    <input
+                      type="number"
+                      min="18"
+                      max="99"
+                      value="18"
+                      class="form-control"
+                      name="edad"
+                      th:field="*{edad}"
+                      placeholder="Escriba la edad del Juez"
+                      required
+                    />
+                  </div>
+
+                  <div class="form-group mb-1">
+                    <label class="control-label" for="email">Email : </label>
+                    <input id="email" type="email" class="form-control" th:field="*{email}" required autofocus="autofocus" />
+                  </div>
+
+                  <div class="form-group">
+                    <label class="control-label" for="password">Password : </label>
+                    <input type="password" id="password" class="form-control" th:field="*{password}" required autofocus="autofocus" />
+                  </div>
+
+                  <br />
+
+                  <button class="btn btn-primary btn-lg btn-block" type="submit">Registrarse</button>
                   <script type="module">
                     import confetti from 'https://cdn.skypack.dev/canvas-confetti';
                     confetti();
@@ -89,8 +120,8 @@
 
                 <hr class="my-4" />
                 <p class="text-center">
-                  ¿Eres entrenador?
-                  <a href="/registro">Registrate aqui</a>
+                  Ya tienes cuenta
+                  <a href="/login">Inicia sesión aquí</a>
                 </p>
               </div>
             </div>
